requirements.txt
twscrape
google-genai
python-telegram-bot
python-dotenv
apscheduler
